Fixed renaming of components
