+++ src/js/components/MainNavButton/MainNavButton.js
@@ -1,50 +1,88 @@
 import React, { Component } from "react";
+import PropTypes from 'prop-types';
 import '../../../scss/components/MainNavButton/MainNavButton.scss';
-import ServerButton from "../ServerButton/ServerButton";
-import AddServerButton from "../AddServerButton/AddServerButton";
 import { GlobalContext } from "../../contexts/GlobalContext";
+
+export const iconClassNames = {
+  svg: 'svg',
+  home: 'svg-home',
+  add: 'svg-add',
+  discover: 'svg-discover',
+  download: 'svg-download',
+}
+
+export const uiClassNames = {
+  base: 'nav-channel',
+  content: 'nav-channel-content',
+  active: 'nav-channel-active',
+  blue: 'nav-channel-blue',
+  green: 'nav-channel-green',
+  image: 'nav-channel-image',
+  separator: 'nav-channel-separator',
+}
 
 class MainNavButton extends Component {
   static contextType = GlobalContext;
-
-  render() {
-    const { mainNavButtons } = this.context;
-
-    return (
-      <div className="nav-column nav-column-server">
-          <div className="nav-group">
-            <ServerButtons buttons={ mainNavButtons } />
-          </div>
-      </div>
-    )
+  static defaultProps = {
+    title: 'Home',
+    iconClassName: null,
+    imageSrc: null,
+    channelExtraClassNames: [],
+    contentExtraClassNames: [],
   }
-}
-
-const ServerButtons = (props) => {
-  const { buttons } = props;
 
   /**
-   * A lookup table of buttons ids -> components
-   *
-   * @todo Move to CMS-able source
+   * Resolve the channel element's class names
+   * @returns {string}
    */
-  const customButtons = [
-    {
-      id: 'add-a-server',
-      component: AddServerButton
-    }
-  ];
+  channelClassNames = () => [
+      uiClassNames.base,
+      ...this.props.channelExtraClassNames,
+      ...(this.props.imageSrc ? [uiClassNames.image] : []),
+      ...(this.isActive() ? [uiClassNames.active] : [])
+    ].join(' ');
 
-  return buttons.map((button, index) => {
-    // Check if the button id is listed in our custom buttons
-    const match = customButtons.find( x => x.id === button.id);
+  /**
+   * Resolve the content element's class names
+   * @returns {string}
+   */
+  contentClassNames = () => [
+    ...(this.props.iconClassName ? [iconClassNames.svg, this.props.iconClassName] : [uiClassNames.content]),
+    ...this.props.contentExtraClassNames,
+  ].join(' ');
 
-    // Use a custom button's component or just default to ServerButton
-    let component = match ? match.component : ServerButton;
+  title = () => this.props.title;
 
-    // Render whichever component is needed
-    return React.createElement(component, { key: index.toString(), ...button });
-  });
-};
+  titleInitials = () => (!(this.props.imageSrc||this.props.iconClassName))
+    ? this.title().split(' ').map((ar)=> ar.charAt(0)).join('')
+    : null;
+
+  isActive = () => {
+    return this.props.id === this.context.currentMainNavButtonId;
+  }
+
+  onClickHandler = () => {
+    this.context.setCurrentMainNavButtonId(this.props.id)
+  }
+
+  render = () => (
+    <button className="nav-item nav-item-server">
+      <div className={ this.channelClassNames() } onClick={ this.onClickHandler } >
+        <div className={ this.contentClassNames() }>{ this.titleInitials() }</div>
+      </div>
+      <div className="pip"/>
+      <div className="tool-tip">{ this.title() }</div>
+    </button>
+  );
+}
+
+MainNavButton.propTypes = {
+  id: PropTypes.string,
+  title: PropTypes.string,
+  iconClassName: PropTypes.string,
+  imageSrc: PropTypes.string,
+  channelExtraClassNames: PropTypes.array,
+  contentExtraClassNames: PropTypes.array,
+}
 
 export default MainNavButton;
