+++ src/scss/components/MainNav/MainNav.scss
@@ -1,0 +1,8 @@
+@import "../../utilities/media";
+@import "../../variables";
+
+.nav-column-server {
+  width: 7.2rem;
+  flex: 0 0 auto;
+  padding-top: 1.2rem;
+}
