--- src/scss/components/ServerButton/ServerButton.scss
@@ -1,167 +1,0 @@
-@import "../../utilities/typography";
-@import "../../variables";
-
-.nav-item-server {
-  align-items: center;
-  justify-content: center;
-  position: relative;
-
-  .pip {
-    position: absolute;
-    top: 14px;
-    left: -4px;
-    width: 8px;
-    height: 2rem;
-    border-radius: 0 4px 4px 0;
-    background-color: $hover-white;
-    transform: scale(0);
-    transition: all 0.2s;
-  }
-
-  .nav-channel {
-    background-color: $neutral-shadow-grey;
-    height: 4.8rem;
-    width: 4.8rem;
-    margin-bottom: 0.8rem;
-    border-radius: 3rem;
-    cursor: pointer;
-    align-items: center;
-    display: flex;
-    justify-content: center;
-    transition: background-color 0.2s, border-radius 0.2s;
-  }
-
-  .tool-tip::before {
-    content: '';
-    position: absolute;
-    top: 0.9rem;
-    left: -0.5rem;
-    width: 0;
-    height: 0;
-    border-top: 10px solid transparent;
-    border-bottom: 10px solid transparent;
-    border-right:10px solid $hover-black;
-  }
-
-  .tool-tip {
-    position: absolute;
-    top: 6px;
-    left: 69px;
-    background-color: $hover-black;
-    border-radius: 4px;
-    padding: 1rem;
-    white-space: nowrap;
-    transform: scale(0);
-    transition: transform 0.015s;
-    @include typo(tt, 1.5em);
-  }
-
-  .nav-channel:hover {
-    border-radius: 1.6rem;
-  }
-
-  .nav-channel-blue:hover {
-    background-color: $hover-blue;
-  }
-
-  .nav-channel-green:hover {
-    background-color: $hover-green;
-  }
-
-  .nav-channel:hover > .svg {
-    filter: brightness(400%);
-  }
-
-  .nav-channel:hover > .nav-channel-content {
-    color: $text-white;
-  }
-
-  .nav-channel:hover ~ .tool-tip {
-    transform: scale(1);
-  }
-
-  .nav-channel:hover ~ .pip {
-    transform: scale(1);
-  }
-
-  .nav-channel-content {
-    color: $light-text-grey;
-    font-size: 1.7em;
-    font-weight: 500;
-    user-select: none;
-    overflow: hidden;
-  }
-
-  .nav-channel-image {
-    background-image: url("../../../img/icons/discord.svg");
-    background-size: contain;
-    background-repeat: no-repeat;
-    background-position: 0 0.5rem;
-  }
-
-  .svg-add {
-    background-image: url("../../../img/icons/plus.svg");
-  }
-  .svg-discover {
-    background-image: url("../../../img/icons/search.svg");
-  }
-  .svg-download {
-    background-image: url("../../../img/icons/download.svg");
-  }
-
-  .svg-home {
-    background-image: url("../../../img/icons/discord.svg");
-    transform: scale(1);
-    height: 2.0rem;
-    width: 2.8rem;
-    filter: brightness(80%);
-    transition: filter 0.2s;
-  }
-
-  .nav-channel:active {
-    transform: translateY(1px);
-  }
-
-  .nav-channel-active {
-    border-radius: 1.6rem;
-  }
-
-  .nav-channel-active.nav-channel-blue {
-    background-color: $normal-brand-blue;
-  }
-
-  .nav-channel-active.nav-channel-green {
-    background-color: $normal-brand-green;
-  }
-
-  .nav-channel-active ~ .pip {
-    transform: scale(1);
-    height: 4rem;
-    top: 0.4rem;
-  }
-
-  .nav-channel-active > .svg {
-    filter: brightness(400%);
-  }
-
-  .nav-channel-active > .nav-channel-content {
-    color: $text-white;
-  }
-
-  .nav-channel-separator {
-    margin-bottom: 1.6rem;
-    position: relative;
-  }
-
-  .nav-channel-separator::after {
-    content: '';
-    position: absolute;
-    bottom: -0.9rem;
-    left: 0.8rem;
-    height: 0.2rem;
-    width: 3.2rem;
-    border-radius: 0.1rem;
-    background-color: $light-outline-grey;
-  }
-}
-
